% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{array_extractor}
\alias{array_extractor}
\title{Extract Parts of an \code{array} Using \code{dimnames}}
\usage{
array_extractor(inarray, valslist)
}
\arguments{
\item{inarray}{The input array.}

\item{valslist}{A list of vectors to use to extract data. A value of \code{NULL}
for any of the list elements will return all values for that dimension.}
}
\value{
An array.
}
\description{
Uses a \code{list} of the same length as the dimensions of an array to extract
information from an array, similar to using matrix indexing.
}
\note{
The \code{list} used for \code{valslist} must be the same length as the number of
dimensions in the array. It must also be specified in the same order as you
would normally reference the dimensions of an array. For instance, in the
example, the array has (in this order) row dimensions, column dimensions, and
a third dimension.
}
\examples{

my_array <- structure(1:12, .Dim = c(2L, 3L, 2L), 
  .Dimnames = list(c("D_11", "D_12"), 
  c("D_21", "D_22", "D_23"), c("D_31", "D_32")))
my_array

array_extractor(my_array, list("D_11", NULL, NULL))
array_extractor(my_array, list(NULL, "D_21", "D_32"))
array_extractor(my_array, list(NULL, c("D_21", "D_22"), NULL))

}
\references{
\url{http://stackoverflow.com/q/34795331/1270695}
}
\author{
Ananda Mahto
}
