% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartesian_split_unlist.R
\name{cartesian_split}
\alias{cartesian_split}
\title{Split Columns Sequentially to Cartesian Product}
\usage{
cartesian_split(indt, splitCols, sep = ",", fixed = TRUE,
  stripWhite = TRUE, type.convert = TRUE)
}
\arguments{
\item{indt}{The input dataset.}

\item{splitCols}{The columns to be split.}

\item{sep}{The separator on which to split each column.}

\item{fixed}{Logical. Should the split character be treated as a fixed pattern
(\code{TRUE}) or a regular expression (\code{FALSE})? Defaults to \code{TRUE}.}

\item{stripWhite}{Logical. Should leading and whitespace be stripped from the
output? Defaults to \code{TRUE}.}

\item{type.convert}{Logical. Should \code{\link[utils:type.convert]{utils::type.convert()}} be used to convert
the result of each column? This might add a little bit of time to the
execution time.}
}
\value{
A \code{data.table}.
}
\description{
Splits each requested column one at a time, resulting in the Cartesian
Product of the split data.
}
\examples{
DC <- data.frame(AB = c("A", "B"), V1 = c("AB,BW", "x,y,z"), V2 = c("1,2,3", "4,5,6,7"))
cartesian_split(DC, c("V1", "V2"))
}
\seealso{
\code{\link[=cSplit]{cSplit()}}
}
\author{
Ananda Mahto
}
