% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartesian_split_unlist.R
\name{cartesian_unlist}
\alias{cartesian_unlist}
\title{Unlist a List Sequentially to Cartesian Product}
\usage{
cartesian_unlist(inlist)
}
\arguments{
\item{inlist}{The input list.}
}
\value{
A \code{data.table}.
}
\description{
Unlists a list sequentially to resulting in the Cartesian Product
of the list elements.
}
\examples{
L2 <- list(X1 = list("A", c("A", "B"), "X", NULL),
           X2 = list(NULL, c(1, 2, 3), c(1, 2), c(1, 2, 3, 4)),
           X3 = list(c("a", "b"), "c", "d", c("e", "f", "g")))
cartesian_unlist(L2)
}
\seealso{
\code{\link[=cartesian_split]{cartesian_split()}}
}
\author{
Ananda Mahto
}
