% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cSplit_e_helper_functions.R
\name{char_mat}
\alias{char_mat}
\title{Create a Binary Matrix from a List of Values}
\usage{
char_mat(listOfValues, mode = "binary", fill = NULL)
}
\arguments{
\item{listOfValues}{A \code{list} of input values to be inserted in a
matrix.}

\item{mode}{Either \code{"binary"}, \code{"value"}, or \code{"count"}. 
Defaults to \code{"binary"}.}

\item{fill}{The initializing fill value for the empty matrix. Defaults to
\code{NA} if \code{mode = "value"} and \code{0} for other modes.}
}
\value{
A \code{matrix}.
}
\description{
Create a binary matrix from a list of character values.
}
\details{
Takes an input \code{list} and converts it to a binary, value, or count 
matrix.
}
\examples{

invec <- c("rock,electro", "electro", "rock,jazz", NA, "jazz,jazz,jazz")
A <- strsplit(invec, ",")
char_mat(A)
char_mat(A, fill = 99)
char_mat(A, mode = "value")
char_mat(A, mode = "count")

}
\seealso{
\code{strsplit}, \code{\link{num_mat}}
}
\author{
Ananda Mahto
}
