% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cSplit_l.R
\name{concat.split.list}
\alias{concat.split.list}
\alias{cSplit_l}
\title{Split Concatenated Cells into a List Format}
\usage{
cSplit_l(indt, splitCols, sep = ",", fixed = TRUE, drop = FALSE,
  stripWhite = FALSE, type.convert = TRUE, ...)
}
\arguments{
\item{indt}{The source \code{data.table}.}

\item{splitCols}{The variables that need to be split. Can be either the name
or the index position.}

\item{sep}{The character or a vector of characters separating each value. Can
also be a regular expression.}

\item{fixed}{Used for \code{\link[base:strsplit]{base::strsplit()}} for allowing regular expressions to
be used when splitting the values.}

\item{drop}{Logical. Should the original variable be dropped? Defaults to
\code{FALSE}.}

\item{stripWhite}{Logical. Should leading and trailing whitespace be trimmed
from the list items? Defaults to \code{FALSE}.}

\item{type.convert}{Logical. Should \code{\link[utils:type.convert]{utils::type.convert()}} be used on the
output? Defaults to \code{TRUE}.}

\item{\dots}{Further arguments to \code{\link[base:strsplit]{base::strsplit()}}.}
}
\value{
A \code{data.table} with the concatenated column split and added as
a \code{list}. If \code{drop = FALSE}, the list-columns would be named in the form of
\code{"original_name_list"}. If \code{drop = TRUE}, the list-columns maintain the
original column names.
}
\description{
Takes a column in a \code{data.table} with multiple values, splits the values into
a \code{list}, and returns a new \code{data.table}.
}
\examples{

DT <- data.frame(ID = 1:5, V1 = c("1,2,3", "2,2,2,4", NA, "", "2,4"), 
                 V2 = c("A ; B", " D; D; D ", "E", "", NA))
cSplit_l(DT, c("V1", "V2"), c(",", ";"))

}
\seealso{
\code{\link[=cSplit]{cSplit()}}, \code{\link[=cSplit_e]{cSplit_e()}}
}
\author{
Ananda Mahto
}
