% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.split.R
\name{concat.split}
\alias{concat.split}
\title{Split Concatenated Cells in a Dataset}
\usage{
concat.split(data, split.col, sep = ",", structure = "compact",
  mode = NULL, type = NULL, drop = FALSE, fixed = FALSE, fill = NA,
  ...)
}
\arguments{
\item{data}{The source \code{data.table}.}

\item{split.col}{The variable that needs to be split; can be specified
either by the column number or the variable name.}

\item{sep}{The character separating each value (defaults to \code{","}.}

\item{structure}{Can be either \code{"compact"}, \code{"expanded"}, or \code{"list"}.
Defaults to \code{"compact"}. See Details.}

\item{mode}{Can be either \code{"binary"}, \code{"value"}, or \code{"count"} (where \code{"binary"}
is default and it recodes values to 1 or 0). This setting only applies when
\code{structure = "expanded"}; a warning message will be issued if used with other
structures.}

\item{type}{Can be either \code{"numeric"} or \code{"character"} (where \code{"numeric"} is
default). This setting only applies when \code{structure = "expanded"}; a warning
message will be issued if used with other structures.}

\item{drop}{Logical. Should the original variable be removed in the output?
Defaults to \code{FALSE}.}

\item{fixed}{Is the input for the \code{sep} value \code{fixed}, or a \emph{regular
expression}? See Details.}

\item{fill}{The "fill" value for missing values when \code{structure = "expanded"}.
Defaults to \code{NA}.}

\item{\dots}{Additional arguments to \code{\link[=cSplit]{cSplit()}}.}
}
\description{
The \code{concat.split} function takes columns with multiple values, splits the
values into a \code{list}, into separate columns, or other specified form, and
returns a new \code{data.table}.
}
\details{
\emph{structure}
\itemize{
\item \code{"compact"} creates as many columns as the maximum length of the resulting
split. This is the most useful general-case application of this function.
\item When the input is numeric, \code{"expanded"} creates as many columns as the
maximum value of the input data. This is most useful when converting to
\code{mode = "binary"}.
\item \code{"list"} creates a single new column that is structurally a \code{list} within a
\code{data.table}.
}

\emph{fixed}
\itemize{
\item When \code{structure = "expanded"} or \code{structure = "list"}, it is possible to
supply a a regular expression containing the characters to split on. For
example, to split on \code{","}, \code{";"}, or \code{"|"}, you can set \code{sep = ",|;|\|"} or
\code{sep = "[,;|]"}, and \code{fixed = FALSE} to split on any of those characters.
}
}
\note{
This is more of a "legacy" or "convenience" wrapper function
encompassing the features available in the separated functions of \code{\link[=cSplit]{cSplit()}},
\code{\link[=cSplit_e]{cSplit_e()}}, and \code{\link[=cSplit_l]{cSplit_l()}}.
}
\examples{

## Load some data
DT <- head(concat.test)

# Split up the second column, selecting by column number
concat.split(DT, 2)

# ... or by name, and drop the offensive first column
concat.split(DT, "Likes", drop = TRUE)

# The "Hates" column uses a different separator
concat.split(DT, "Hates", sep = ";", drop = TRUE)

\dontrun{
# You'll get a warning here, when trying to retain the original values
concat.split(DT, 2, mode = "value", drop = TRUE)
}

# Try again. Notice the differing number of resulting columns
concat.split(DT, 2, structure = "expanded", mode = "value", 
             type = "numeric", drop = TRUE)

# Let's try splitting some strings... Same syntax
concat.split(DT, 3, drop = TRUE)

# Strings can also be split to binary representations
concat.split(DT, 3, structure = "expanded", mode = "binary",
             type = "character", fill = 0, drop = TRUE)

# Split up the "Likes column" into a list variable; retain original column
head(concat.split(DT, 2, structure = "list", drop = FALSE))

# View the structure of the output to verify that the new column is a list; 
# note the difference between "Likes" and "Likes_list".
str(concat.split(DT, 2, structure = "list", drop = FALSE))

}
\author{
Ananda Mahto
}
