% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concat.test}
\alias{concat.test}
\title{Example Dataset with Concatenated Cells}
\format{A \code{data.frame} in which many columns contain concatenated cells.}
\description{
This is a sample dataset to demonstrate the different features of the
\code{\link[=concat.split]{concat.split()}} family of functions.
}
\details{
The \code{concat.test} dataset contains three columns with concatenated
data:
\itemize{
\item \code{"Likes"}: Comma-separated integer values, with no spaces. The full value
range for the column would be 1 to 6. Some values are blank (\code{""}) and others
are \code{NA}.
\item \code{"Siblings"}: Comma-separated string values, with spaces before and after
the delimiter. Some values end with the delimiter; others don't.
\item \code{"Hates"}: Semicolon-separated integer values, with no spaces. The full
value range for the column would be 1 to 4. Some values are blank (\code{""}) and
others are \code{NA}. If not blank or \code{NA}, the value ends with a trailing
semicolon.
}
}
\keyword{datasets}
