% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cSplit.R
\name{cSplit}
\alias{cSplit}
\title{Split Concatenated Values in Columns into Separate Columns}
\usage{
cSplit(indt, splitCols, sep = ",", direction = "wide", fixed = TRUE,
  drop = TRUE, stripWhite = TRUE, makeEqual = NULL,
  type.convert = TRUE)
}
\arguments{
\item{indt}{The input \code{data.table}.}

\item{splitCols}{The column or columns that have values which need to be split.}

\item{sep}{The values that serve as a delimiter \emph{within} each column. This
can be a single value if all columns have the same delimiter, or a vector of
values \emph{in the same order as the delimiters in each of the \code{splitCols}}.}

\item{direction}{The desired direction of the results, either \code{"wide"} or \code{"long"}.}

\item{fixed}{Logical. Should the split character be treated as a fixed
pattern (\code{TRUE}) or a regular expression (\code{FALSE})? Defaults to \code{TRUE}.}

\item{drop}{Logical. Should the original concatenated column be dropped?
Defaults to \code{TRUE}.}

\item{stripWhite}{Logical. If there is whitespace around the delimiter in
the concatenated columns, should it be stripped from the resulting columns?
Defaults to \code{TRUE}.}

\item{makeEqual}{Logical. Should all groups be made to be the same length or
the same number of columns? Defaults to \code{FALSE}.}

\item{type.convert}{Logical. Should \code{\link[utils:type.convert]{utils::type.convert()}} be used to convert
the result of each column? This might add a little bit of time to the
execution time.}
}
\value{
A \code{data.table} with the values split into new columns or rows.
}
\description{
The \code{cSplit} function is designed to quickly and conveniently split
concatenated data into separate columns.
}
\note{
The \code{cSplit} function replaces most of the earlier \code{concat.split*}
functions. The earlier functions remain for compatability purposes, but now
they are essentially wrappers around the \code{cSplit} function and are likely to
be deprecated over time.
}
\examples{

DT <- head(concat.test)

## Split the "Likes" column
cSplit(DT, "Likes")

## Split the "Likes" and "Hates" columns --
##   they have different delimiters...
cSplit(DT, c("Likes", "Hates"), c(",", ";"))

## Split "Siblings" into a long form...
cSplit(DT, "Siblings", ",", direction = "long")

## Split "Siblings" into a long form, keeping extra whitespace
cSplit(DT, "Siblings", ",", direction = "long", stripWhite = FALSE)

## Split a vector
y <- c("a_b_c", "a_b", "c_a_b")
cSplit(data.frame(y), "y", "_")

}
\seealso{
\code{\link[=concat.split]{concat.split()}}
}
\author{
Ananda Mahto
}
