% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cSplit_f.R
\name{cSplit_f}
\alias{cSplit_f}
\title{Split Concatenated Cells in a data.frame or a data.table}
\usage{
cSplit_f(indt, splitCols, sep, drop = TRUE, dotsub = "|",
  stripWhite = FALSE)
}
\arguments{
\item{indt}{The input \code{data.frame} or \code{data.table}.}

\item{splitCols}{The columns that need to be split up.}

\item{sep}{The character or characters that serve as delimiters within the
columns that need to be split up. If different columns use different
delimiters, enter the delimiters as a character vector.}

\item{drop}{Logical. Should the original columns be dropped? Defaults to
\code{TRUE}.}

\item{dotsub}{The character that should be substituted as a delimiter *if
\code{sep = "."}. \code{fread} does not seem to always work nicely with \code{sep = "."}, so
it needs to be substituted. By default, this function will substitute \code{"."}
with \code{"|"}.}

\item{stripWhite}{Logical. Should whitespace be stripped before writing to
the temporary file? Defaults to \code{FALSE}.}
}
\value{
A \code{data.table}.
}
\description{
A variation of the \code{\link[=concat.split]{concat.split()}} family of functions that uses
\code{\link[data.table:fread]{data.table::fread()}} for speedy splitting of concatenated columns of data.
}
\examples{

## Sample data. Change `n` to larger values to test on larger data
set.seed(1)
n <- 10
mydf <- data.frame(id = sequence(n))
mydf <- within(mydf, {
  v3 <- do.call(paste, c(data.frame(matrix(sample(
  letters, n*4, TRUE), ncol = 4)), sep = "_"))
  v2 <- do.call(paste, c(data.frame(matrix(sample(
  LETTERS, n*3, TRUE), ncol = 3)), sep = "."))
  v1 <- do.call(paste, c(data.frame(matrix(sample(
  10, n*2, TRUE), ncol = 2)), sep = "-"))
})
mydf

cSplit_f(mydf, splitCols = c("v1", "v2", "v3"), sep = c("-", ".", "_"))

}
\references{
\url{http://stackoverflow.com/a/19231054/1270695}
}
\author{
Ananda Mahto. Thanks also to Arun Srinivasan for helping to refine
this function.
}
