% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitstackshape_helper_functions.R
\name{dataset_names}
\alias{dataset_names}
\alias{othernames}
\alias{Names}
\title{Name Convenience Functions}
\usage{
othernames(data, toremove)

Names(data, invec)
}
\arguments{
\item{data}{The input dataset.}

\item{toremove}{The \code{\link[base:names]{base::names()}} you want to exclude.}

\item{invec}{The \code{\link[base:names]{base::names()}} you want to keep.}
}
\value{
A character vector of the remaining names.
}
\description{
\code{othernames} is a convenience function for
\code{setdiff(names(data), -some_vector_of_names-)}. \code{Names} is a convenience
function using either character vectors or numeric vectors to specify a
subset of \code{\link[base:names]{base::names()}} of a dataset, returning the names as a character
vector, always.
}
\examples{

mydf <- data.frame(a = 1:2, b = 3:4, c = 5:6)
splitstackshape:::othernames(mydf, "a")


mydf <- data.frame(a = 1:2, b = 3:4, c = 5:6)
splitstackshape:::Names(mydf, c("a", "c"))
splitstackshape:::Names(mydf, c(1, 3))

}
\seealso{
\code{\link[base:setdiff]{base::setdiff()}}
}
\author{
Ananda Mahto
}
