% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2dt.R
\name{dist2dt}
\alias{dist2dt}
\title{Converts a Distance Matrix to a data.table}
\usage{
dist2dt(inDist)
}
\arguments{
\item{inDist}{The input distance object.}
}
\value{
A \code{data.table}.
}
\description{
Converts a distance matrix to a \code{data.table}.
}
\examples{

dd <- as.dist((1 - cor(USJudgeRatings)[1:5, 1:5])/2)
dist2dt(dd)

}
\references{
\url{http://stackoverflow.com/q/23474729/1270695}
}
\author{
Ananda Mahto
}
