% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_long.R
\name{merged.stack}
\alias{merged.stack}
\title{Reshapes Wide Data According to Variable Stub Values}
\usage{
merged.stack(indt, id.vars = NULL, var.stubs, keep.all = TRUE,
  end_stub = FALSE, sep = NULL)
}
\arguments{
\item{indt}{The input \code{data.table}.}

\item{id.vars}{The columns to be used as "ID" variables. Defaults to \code{NULL},
at which point, all names which are not identified as variable groups are
used as the identifiers.}

\item{var.stubs}{The prefixes of the variable groups.}

\item{keep.all}{Logical. Should all the variables in the source data be kept
(\code{keep.all = TRUE}) or only those which comprise the \code{id.vars} and split data
from the \code{var.stubs} (\code{keep.all = FALSE}).}

\item{end_stub}{Logical. Is the stub at the start of the variable name (eg:
"var.1") or at the end of the variable name ("1.var")? Defaults to \code{FALSE}
indicating that a naming pattern such as "var.1" is expected.}

\item{sep}{Legacy argument. No longer implemented.}
}
\value{
A \code{data.table}.
}
\description{
A wrapper around the \code{\link[=reshape_long]{reshape_long()}} function. Will be removed from future
versions of "splitstackshape".
}
\examples{

DF <- data.frame(id_1 = 1:3, id_2 = c("A", "B", "A"),
                 varA.1 = c("a", "b", "c"), varA.2 = c("x", "y", "z"), 
                 varA.3 = c("q", "r", "s"), varB.2 = c(10, 20, 30),
                 varB.3 = c(11, 22, 33), varC.3 = c(-1.23, .992, -.351))
merged.stack(DF, var.stubs = c("varA", "varB", "varC"))

}
\seealso{
\code{\link[=reshape_long]{reshape_long()}}
}
\author{
Ananda Mahto
}
