% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_me.R
\name{move_me}
\alias{move_me}
\title{Reorders the Contents of a Vector}
\usage{
move_me(invec, movecommand)
}
\arguments{
\item{invec}{The input vector}

\item{movecommand}{The command that describes how you want to shuffle the
vector. See \emph{Details}.}
}
\value{
A vector.
}
\description{
Shuffle the order of a vector around using natural language statements.
}
\details{
This can be a useful function for reordering the columns of a \code{data.table} in
a convenient manner. In such cases, the \code{invec} would be \code{names(your_data)}.

The \code{movecommand} argument is specified in the form of \code{"a, b before f"}. The positions to move are:
\itemize{
\item \strong{\code{first}}: move the specified items to the first postion.
\item \strong{\code{last}}: move the specified items to the last position.
\item \strong{\code{before}}: move the specified items before the value mentioned.
\item \strong{\code{after}}: move the specified items after the value mentioned.
}

Multiples are allowed:
\itemize{
\item Specify multiple values to be moved by separating them with a comma.
\item Chain multiple move commands by separating them with a semicolon.
}
}
\examples{

myvec <- letters[1:10]
myvec
move_me(myvec, "a last; b, e, g before d; c first; h after j")

new_col_order <- move_me(
  names(mtcars), "hp first; cyl after drat; vs, am, gear before mpg; wt last")
mtcars[new_col_order]

}
\references{
\url{http://stackoverflow.com/a/18420673/1270695}
}
\author{
Ananda Mahto
}
