% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cSplit_e_helper_functions.R
\name{num_mat}
\alias{num_mat}
\title{Create a Numeric Matrix from a List of Values}
\usage{
num_mat(listOfValues, mode = "binary", fill = NULL)
}
\arguments{
\item{listOfValues}{A \code{list} of input values to be inserted in a
matrix.}

\item{mode}{Either \code{"binary"}, \code{"value"}, or \code{"count"}. 
Defaults to \code{"binary"}.}

\item{fill}{The initializing fill value for the empty matrix. Defaults to
\code{NA} if \code{mode = "value"} and \code{0} for other modes.}
}
\value{
A \code{matrix}.
}
\description{
Create a numeric matrix from a list of values.
}
\details{
Takes an input \code{list} and converts it to a binary, value, or count 
matrix.
}
\examples{

invec <- c("1,2,4,5,6", "1,2,4,5,6", NA,
           "1,2,4,5,6,6,6", "-1,1,2,5,6", "1,2,5,6")
A <- strsplit(invec, ",")
num_mat(A)
num_mat(A, fill = 0)
num_mat(A, mode = "value")
num_mat(A, mode = "count")

}
\seealso{
\code{strsplit}, \code{\link{char_mat}}
}
\author{
Ananda Mahto
}
