% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitstackshape_helper_functions.R
\name{othernames}
\alias{othernames}
\title{Extract All Names From a Dataset Other Than the Ones Listed}
\usage{
othernames(data, toremove)
}
\arguments{
\item{data}{The input \code{data.frame}.}

\item{toremove}{The \code{names} you want to exclude.}
}
\value{
A character vector of the remaining names.
}
\description{
A convenience function for \code{setdiff(names(data), -some_vector_of_names-)}.
}
\examples{

mydf <- data.frame(a = 1:2, b = 3:4, c = 5:6)
splitstackshape:::othernames(mydf, "a")

}
\seealso{
\code{\link[base:setdiff]{base::setdiff()}}
}
\author{
Ananda Mahto
}
