% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_long.R
\name{reshape_long}
\alias{reshape_long}
\title{Reshape Unbalanced Wide Data Into a Long Form}
\usage{
reshape_long(indt, stubs, ids = NULL, value.name = NULL,
  variable.name = NULL, end_stub = FALSE, keep.all = TRUE)
}
\arguments{
\item{indt}{The input \code{data.table}.}

\item{stubs}{The prefixes or suffixes of the varying groups. This can be a
\code{name}d character vector. See "Details".}

\item{ids}{The columns to be used as "ID" variables. Defaults to \code{NULL} at
which point, all names which are not identified as part of the varying
columns are used.}

\item{value.name}{An alternative name to be used for the resulting
consolidated columns. See "Details".}

\item{variable.name}{An alternative name to be used instead of "variable" to
indicate the measurement or observation period}

\item{end_stub}{Logical. Is the stub at the start of the variable name (eg:
"var.1") or at the end of the variable name ("1.var")? Defaults to \code{FALSE}
indicating that a naming pattern such as "var.1" is expected.}

\item{keep.all}{Logical. Should all columns be kept, or only those specified
in the combination of column names from "stubs" and "ids"? Defaults to \code{TRUE}.}
}
\value{
A \code{data.table}.
}
\description{
Takes unbalanced wide data (where measurements may not all have
the same number of time or observation variables) and converts it into a long
format.
}
\details{
The names of the resulting columns can be provided in a few different ways.
\itemize{
\item If nothing is specified, the provided "stubs" are used as the new column
names.
\item If the "stubs" value is a named vector, the names of each vector element
is used for the new column names.
\item If "variable.name" is specified, the names provided there will be used for
the new column names.
\item If "stubs" is a named vector \emph{and} "variable.name" has been specified,
names are taken from the names in the "stubs" vector and those from
"variable.name" are ignored.
}
}
\examples{

DF <- data.frame(id_1 = 1:3, id_2 = c("A", "B", "A"),
                 varA.1 = c("a", "b", "c"), varA.2 = c("x", "y", "z"), 
                 varA.3 = c("q", "r", "s"), varB.2 = c(10, 20, 30),
                 varB.3 = c(11, 22, 33), varC.3 = c(-1.23, .992, -.351))
reshape_long(DF, c("varA", "varB", "varC"))
reshape_long(DF, c(colA = "varA", colB = "varB", colC = "varC"))
 
}
\author{
Ananda Mahto
}
