% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_splitters.R
\name{t_split}
\alias{t_split}
\alias{t_split}
\alias{f_split}
\title{Split Concatenated Vectors into Multi-column \code{data.table}s}
\usage{
t_split(vec, sep, fixed = TRUE, stripWhite = TRUE, type.convert = TRUE,
  prefix = NULL)

f_split(vec, sep, fixed = TRUE, stripWhite = TRUE, type.convert = TRUE,
  prefix = NULL)
}
\arguments{
\item{vec}{The input vector.}

\item{sep}{The character or regular expression on which to split each value.}

\item{fixed}{Logical. Is the \code{sep} value fixed or a regular expression?
Defaults to \code{TRUE}.}

\item{stripWhite}{Logical. Should leading and trailing whitespace be stripped
from the output? Defaults to \code{TRUE}.}

\item{type.convert}{Logical. Should the output be converted using the
\code{\link[utils:type.convert]{utils::type.convert()}} function? Defaults to \code{TRUE}.}

\item{prefix}{Single character string indicating the desired column prefix
for the resulting column names. Defaults to \code{NULL} at which point names are
simply \code{V1} to \code{Vn}. If a prefix is supplied the resulting names are in the
form of \code{prefix_column_number}.}
}
\value{
A \code{data.table}.
}
\description{
Takes an input of a single vector and splits on a delimiter or
a regular expression to create multiple columns.
}
\note{
\code{t_split()} and \code{f_split()} may return different number of columns if a
\code{sep} value is found at the end of a string being split.

\code{t_split()} is automatically called with a \code{message} if:
\itemize{
\item the result of \code{f_split()} is a single column.
\item \code{fixed = FALSE}.
\item the \code{sep} provided is more than 1 character long, or is equal to \code{"."} or \code{""}.
}
}
\examples{

vec <- c("b;c;d", "d;ef", "d;ef", "", NA, "x;yz")
t_split(vec, ";")
t_split(vec, ";", prefix = "new")


vec <- c("rock,electro", "rock,jazz", NA, "jazz,jazz,jazz")
f_split(vec, ",")
f_split(vec, ",", prefix = "new")

# Expect messages on the following
vec <- c("rockSEPelectro", "rockSEPjazz", NA, "jazzSEPjazzSEPjazz")
f_split(vec, "SEP")
f_split(vec, "[A-Z]+", fixed = FALSE)

}
\author{
Ananda Mahto
}
