% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cSplit_e_helper_functions.R
\name{trim_list}
\alias{trim_list}
\title{Trim Whitespace from a List of Character Vectors}
\usage{
trim_list(x, relist = TRUE, convert = FALSE)
}
\arguments{
\item{x}{A \code{list} of character vectors.}

\item{relist}{Logical. Should the output also be a \code{list}? Defaults to
\code{TRUE}.}

\item{convert}{Logical. Should \code{type.convert} be used on the output?
Defaults to \code{FALSE}.}
}
\value{
A \code{list} or the unlisted values.
}
\description{
Trims whitespace from a list of character vectors.
}
\details{
Takes an input \code{list} of character vectors trims any leading or
trailing whitespace, and optionally relists the output.
}
\note{
Empty list elements or empty characters (eg: \code{""}) are converted
to \code{NA}.
}
\examples{

L1 <- list(c(" A", "B   ", NA), NA, c("   A", NA, ""), "", "B   ")
trim_list(L1)
trim_list(L1, FALSE)

L2 <- list(c(" 1", "2   ", NA), NA, c("   1", NA, ""), "", "3   ")
trim_list(L2)
trim_list(L2, convert = TRUE)

}
\seealso{
\code{trimws}
}
\author{
Ananda Mahto
}
