% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cSplit_e_helper_functions.R
\name{trim_vec}
\alias{trim_vec}
\title{Trim Whitespace from a Character Vector}
\usage{
trim_vec(vec, attr = FALSE)
}
\arguments{
\item{vec}{A character vector.}

\item{attr}{Logical. Should an attribute be set indicating that the the 
output was "clean" or has been "cleaned"? Defaults to \code{FALSE}.}
}
\value{
A vector of the trimmed values.
}
\description{
Trims whitespace from a character vector.
}
\details{
Takes an input character vector and trims any leading or trailing whitespace.
}
\note{
The \code{attr} argument is used to increase the efficiency of
\code{\link{trim_list}}.
}
\examples{

V1 <- letters
trim_vec(V1)
trim_vec(V1, TRUE)

V2 <- c(" 1", "2   ", NA, "")
trim_vec(V2)
trim_vec(V2, TRUE)

}
\seealso{
\code{trimws}, \code{\link{trim_list}}
}
\author{
Ananda Mahto
}
