% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_cols.R
\name{unlist_cols}
\alias{unlist_cols}
\alias{listCol_l}
\alias{listCol_w}
\alias{unlist_cols}
\title{Unlists or Flattens List Columns in a \code{data.table}}
\usage{
listCol_l(indt, listCols, drop = TRUE, makeEqual = FALSE)

listCol_w(indt, listCols, drop = TRUE, makeEqual = FALSE)

unlist_cols(indt, listCols, direction = "wide", drop = TRUE,
  makeEqual = FALSE)
}
\arguments{
\item{indt}{The input \code{data.table}.}

\item{listCols}{The columnst to unlist or flatten.}

\item{drop}{Logical. Should the original column be dropped? Defaults to \code{TRUE}.}

\item{makeEqual}{Logical. Should the lengths of values be made equal before
converting to the long form?}

\item{direction}{Use \code{"long"} for unlisting behavior, and \code{"wide"} for
flattening behavior.}
}
\value{
A \code{data.table}.
}
\description{
Unlists columns stored as \code{list}s into a long form or flattens
them into a wide form.
}
\examples{

dat <- data.frame(A = 1:3, B = I(list(c(1, 2), c(1, 3, 5), c(4))))
listCol_l(dat, "B")


dat <- data.frame(A = 1:3, B = I(list(c(1, 2), c(1, 3, 5), c(4))))
listCol_w(dat, "B")


dat <- data.frame(A = 1:3, B = I(list(c(1, 2), c(1, 3, 5), c(4))))
unlist_cols(dat, "B", "wide")
unlist_cols(dat, "B", "long")

}
\author{
Ananda Mahto
}
