% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.split.R
\name{concat.split.expanded}
\alias{concat.split.expanded}
\alias{cSplit_e}
\title{Split Concatenated Values into their Corresponding Column Position}
\usage{
cSplit_e(indt, splitCols, sep = ",", mode = "binary", type = "numeric",
  drop = FALSE, fixed = TRUE, fill = NULL)
}
\arguments{
\item{indt}{The source \code{data.table}.}

\item{splitCols}{The variables that need to be split. Can be either the name
or the index position.}

\item{sep}{The character or a vector of characters separating each value. Can
also be a regular expression.}

\item{mode}{Can be either \code{"binary"} (where the presence of a number in a
given column is converted to "1"), \code{"value"} (where the value is retained and
not recoded to "1"), or \code{"count"} (where the frequency of each value is
is returned). Defaults to \code{"binary"}.}

\item{type}{Can be either \code{"numeric"} (where the items being split are
integers) or \code{"character"} (where the items being split are character
strings). Defaults to \code{"numeric"}.}

\item{drop}{Logical. Should the original variable be dropped? Defaults to
\code{FALSE}.}

\item{fixed}{Used for \code{\link[base:strsplit]{base::strsplit()}} for allowing regular expressions to
be used when splitting the values.}

\item{fill}{Desired "fill" value. Defaults to \code{NULL} at which reasonable
options are provided by default. If \code{mode = "binary"} or \code{mode = "count"},
the default fill would be "0". If \code{mode = "value"}, the default fill would
be \code{NA}.}
}
\value{
A \code{data.table}.
}
\description{
"Expand" concatenated numeric or character values to their relevant position
in a \code{data.table} or create a binary or count representation of such data.
}
\examples{

temp <- head(concat.test)
cSplit_e(temp, "Likes")
cSplit_e(temp, 4, ";", fill = 999)
 
## The old function name still works
concat.split.expanded(temp, "Likes")
concat.split.expanded(temp, 4, ";", fill = 0)
concat.split.expanded(temp, 4, ";", mode = "value", drop = TRUE)
concat.split.expanded(temp, "Siblings", type = "character", drop = TRUE)

}
\seealso{
\code{\link[=concat.split]{concat.split()}}, \code{\link[=concat.split.list]{concat.split.list()}}, \code{\link[=num_mat]{num_mat()}}, \code{\link[=char_mat]{char_mat()}}
}
\author{
Ananda Mahto
}
