\docType{data}
\name{concat.test}
\alias{concatenated}
\alias{concat.test}
\title{Example dataset with concatenated cells}
\format{A \code{data.frame} in which many columns contain concatenated cells}
\description{
  This is a sample dataset to demonstrate the different
  features of the \code{\link{concat.split}} family of
  functions.
}
\keyword{datasets}

