% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{concat.test}
\alias{concat.test}
\alias{concatenated}
\title{Example Dataset with Concatenated Cells}
\format{A \code{data.frame} in which many columns contain concatenated cells}
\description{
This is a sample dataset to demonstrate the different features of the
\code{\link{concat.split}} family of functions.
}
\keyword{datasets}

