% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cSplit_f.R
\name{cSplit_f}
\alias{cSplit_f}
\title{Split Concatenated Cells in a data.frame or a data.table}
\usage{
cSplit_f(indt, splitCols, sep, drop = TRUE, stripWhite = TRUE, type.convert)
}
\arguments{
\item{indt}{The input \code{data.frame} or \code{data.table}.}

\item{splitCols}{The columns that need to be split up.}

\item{sep}{The character or characters that serve as delimiters within the
columns that need to be split up. If different columns use different
delimiters, enter the delimiters as a character vector.}

\item{drop}{Logical. Should the original columns be dropped? Defaults to
\code{TRUE}.}

\item{stripWhite}{Logical. Should whitespace be removed from character
columns? Defaults to \code{TRUE}.}

\item{type.convert}{Logical. Should the output be converted to relevant types
or kept as character columns? Defaults to \code{TRUE}.}
}
\value{
A \code{data.table}.
}
\description{
Deprecated. Simply calls \code{\link{cSplit}}, so it is better to use that
function directly rather than continue using this one.
}
\examples{

## Sample data. Change `n` to larger values to test on larger data
set.seed(1)
n <- 10
mydf <- data.frame(id = sequence(n))
mydf <- within(mydf, {
  v3 <- do.call(paste, c(data.frame(matrix(sample(
  letters, n*4, TRUE), ncol = 4)), sep = "_"))
  v2 <- do.call(paste, c(data.frame(matrix(sample(
  LETTERS, n*3, TRUE), ncol = 3)), sep = "."))
  v1 <- do.call(paste, c(data.frame(matrix(sample(
  10, n*2, TRUE), ncol = 2)), sep = "-"))
})
mydf

cSplit_f(mydf, splitCols = c("v1", "v2", "v3"), sep = c("-", ".", "_"))

}
\author{
Ananda Mahto.
}
