% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandRows.R
\name{expandRows}
\alias{expandRows}
\title{Expand the Rows of a Dataset}
\usage{
expandRows(indt, count, count.is.col = TRUE, drop = TRUE)
}
\arguments{
\item{indt}{The input \code{data.table}.}

\item{count}{The numeric vector of counts \emph{OR} the column from the dataset
that contains the count data. If \code{count} is a single digit, it is assumed
that all rows should be repeated by this amount.}

\item{count.is.col}{Logical. Is the \code{count} value a column from the input
dataset? Defaults to \code{TRUE}.}

\item{drop}{Logical. If \code{count.is.col = TRUE}, should the "count" column be
dropped from the result? Defaults to \code{TRUE}.}
}
\value{
A \code{data.table}.
}
\description{
Expands (replicates) the rows of a \code{data.table}, either by a fixed number, a
specified vector, or a value contained in one of the columns in the source
\code{data.table}.
}
\examples{

mydf <- data.frame(x = c("a", "b", "q"), y = c("c", "d", "r"), count = c(2, 5, 3))
expandRows(mydf, "count")
expandRows(mydf, count = 3) ## This takes values from the third column!
expandRows(mydf, count = 3, count.is.col = FALSE)
expandRows(mydf, count = c(1, 5, 9), count.is.col = FALSE)

}
\author{
Ananda Mahto
}
