% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_ftable}
\alias{extract_ftable}
\alias{[.ftable}
\title{Extract Parts of an \code{ftable}.}
\usage{
\method{[}{ftable}(inftable, ...)
}
\arguments{
\item{inftable}{The input \code{ftable}.}
}
\description{
Extension of \code{?Extract} to work with \code{\link[stats:ftable]{stats::ftable()}} by using
the names of the columns and rows (stored in the attributes of the \code{ftable}).
}
\examples{
mytable <- ftable(Titanic, row.vars = 1:3)
mytable[c("1st", "3rd"), , , "No"]
tab2 <- ftable(Titanic, row.vars = 1:2, col.vars = 3:4)
tab2[c("1st", "3rd"), , , ]
}
