% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitstackshape_helper_functions.R
\name{FacsToChars}
\alias{FacsToChars}
\title{Convert All Factor Columns to Character Columns}
\usage{
FacsToChars(mydf)
}
\arguments{
\item{mydf}{The name of your \code{data.frame}}
}
\description{
Sometimes, we forget to use the \code{stringsAsFactors} argument when using
\code{\link[utils:read.table]{utils::read.table()}} and related functions. By default, R converts character
columns to factors. Instead of re-reading the data, the \code{FacsToChars}
function will identify which columns are currently factors, and convert them
all to characters.
}
\examples{

## Some example data
dat <- data.frame(title = c("title1", "title2", "title3"),
         author = c("author1", "author2", "author3"),
         customerID = c(1, 2, 1))

str(dat) # current structure
dat2 <- splitstackshape:::FacsToChars(dat)
str(dat2) # Your new object
str(dat)  # Original object is unaffected

}
\seealso{
\code{\link[utils:read.table]{utils::read.table()}}
}
\author{
Ananda Mahto
}
