% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getanID.R
\name{getanID}
\alias{getanID}
\title{Add an "id" Variable to a Dataset}
\usage{
getanID(indt, id.vars = NULL)
}
\arguments{
\item{id.vars}{The variables that should be treated as ID variables. Defaults
to \code{NULL}, at which point all variables are used to create the new ID variable.}

\item{data}{The input \code{data.table}.}
}
\value{
The input dataset if ID variables are already unique, or the input
dataset with a new column named "\code{sub_id}".
}
\description{
Many functions will not work properly if there are duplicated ID variables in
a dataset. This function is a convenience function for \code{\link[data.table:rowid]{data.table::rowid()}}
and is largely unnecessary now that \code{rowid} exists.
}
\examples{

mydf <- data.frame(IDA = c("a", "a", "a", "b", "b"),
                   IDB = c(1, 1, 1, 1, 1), values = 1:5)
mydf
getanID(mydf, c("IDA", "IDB"))

mydf <- data.frame(IDA = c("a", "a", "a", "b", "b"),
                   IDB = c(1, 2, 1, 1, 2), values = 1:5)
mydf
getanID(mydf, 1:2)

}
\author{
Ananda Mahto
}
