% Generated by roxygen2 (4.0.2): do not edit by hand
\name{listCol_l}
\alias{listCol_l}
\title{Unlist a Column Stored as a List}
\usage{
listCol_l(inDT, listcol, drop = TRUE)
}
\arguments{
\item{inDT}{The input dataset.}

\item{listcol}{The name of the column stored as a \code{list}.}

\item{drop}{Logical. Should the original column be dropped? Defaults to \code{TRUE}.}
}
\value{
A \code{data.table}.
}
\description{
Unlists a column stored as a \code{list} into a long form.
}
\examples{
dat <- data.frame(A = 1:3, B = I(list(c(1, 2), c(1, 3, 5), c(4))))
listCol_l(dat, "B")
}
\author{
Ananda Mahto
}
\seealso{
\code{\link{listCol_w}} to flatten a \code{list} column into a "wide" format.
}

